catch errors when encoding entity map
Summary:
We've had several reports of people's content failing to save in the Performance tool. The only error message we have is this (with no callstack, unfortunately):
```
Unknown DraftEntity key: undefined.
```
I poked around in the DraftJS code, and my guess is that this is due to some character having an entity key associated with it that for whatever reason does not appear in the entity map. This diff "fixes" the issue by not throwing an error in this situation.

This is definitely not as good as figuring out the root cause of the entity mismatch, but it seems reasonable for `convertFromDraftStateToRaw` to work with slightly malformed input.

Not sure if it's ok to use `FBLogger` in this code or not, but I saw another usage of it in `draft-js-contrib` here:
diffusion/WWW/browse/master/html/shared/draft-js-contrib/matchers/getImplicitDateTimeMatches.js;1000956292$34

I could also just ignore the error, but the linter didn't like that ¯\_(ツ)_/¯

Reviewed By: niveditc

Differential Revision: D16362778

fbshipit-source-id: 6e2a6041d1a02b412d94655c9b6224fee868748d