Flow-type DataTransfer.js
Summary:
Flowifies DataTransfer.js. We have a flow libdef for the browser's DataTransfer object, **so the only non-flow change this involved was a quick name check to avoid the collision**.

This surfaced a couple of places where the module is misused (yay! means it's surfacing potential bugs!). Putting them behind flowfixme's to avoid changing functionality and keep this diff simple (and eye-testable).

I can make tasks for fixing these flowfixme's if desired.

Differential Revision: D16176363

fbshipit-source-id: 755da606d0d86b74f2024b6d8e522746072bfc61