Fix isHTMLBRElement test (#2278)
Summary:
**Summary**

In the Facebook repo mirror, modules in Jest tests are not mocked, but on GitHub they are (as seen from `package.json`). That led to D18831442 passing internally in Sandcastle but failing on Travis. Not a huge deal, we can just add `jest.disableAutomock()`.

We might want to see if we can remove `"automock": true,` from the Jest config in `package.json` to have a more similar environment to Facebook's Jest config.

**Test Plan**

Travis CI should pass now.
Pull Request resolved: https://github.com/facebook/draft-js/pull/2278

Differential Revision: D18874041

Pulled By: yangshun

fbshipit-source-id: 282c137934dea27259a2b52741c749fb6c332323