IE could not display composer when opening or creating a new group chat from the chat create view
Summary: When opening a group chat from workplace chat's new message in IE the first invocation of interfaceselection.addRange throws an unspecified error. Remarkably, subsequent invocations work fine and draft editor is able to continue as usual. This typically happens when IE doesn't like something but it is very hard to suss out what the root cause is. So I'm wrapping with a try catch so that this does not crash the draft editor completely.

Reviewed By: danielbuechele

Differential Revision: D19163859

fbshipit-source-id: ded92cae390dfca8d8956a49c1be54b1540b2998